Working tile layers, build changes
Tile layers 1 and 2 now display correctly, thanks to the correction of an offset in the _fetch_data() function.
Sprites are still corrupted though.

libcore and libui have been dropped in favor of standard per-file objects. This avoids PIC, and it seems like a 20-30% performance boost is obtained.

Formatting and logging changes.